Reimplemented putting SQLite filedb in WAL mode.
SQLAlchemy v2 somehow causes an sqlite disk I/0 error when creating tables when DB is in WAL mode. So now we put it in WAL mode before any transaction is opened, which hopefully will mean tables will be created before database is put into WAL mode.

This might have something to do with autocommit mode changes. If so, can revert to previous behaviour of setting WAL mode when Datastore class is constructed, and find a way to avoid autocommit when creating tables?