Improved transactions, so they are effectively thread-local scoped.
This was done because although passing a session object down the stack from the save method the session works okay, command methods that also need to get an aggregate before changing and then saving it won't work in SQLite because the access is locked, and in any case it would be better to do this in the same session, and it's not possible to pass the session to the repository get() method. This works okay, but it would be better to use a scoped_session object, so that request-scoped sessions from integrations between web frameworks and SQLAlchemy that provide can be used.

Also reworked the SQLDatabase class to initialise in things more clearly.

Also added 'SQLALCHEMY_AUTOFLUSH' configuration.